Created new modules file
Moved the modules from within the script to a separate my_modules.py file
